+++ packages/cli/src/ui/hooks/useLogger.ts
@@ -17,7 +17,7 @@ export const useLogger = () => {
     const newLogger = new Logger(sessionId());
     /**
      * Start async initialization, no need to await. Using await slows down the
-     * time from launch to see the opencli prompt and it's better to not save
+     * time from launch to see the grokcli prompt and it's better to not save
      * messages than for the cli to hanging waiting for the logger to loading.
      */
     newLogger